requirements.txt
# Python 3.10.18
tensorflow-macos 
tensorflow-metal
"gymnasium[atari,accept-rom-license]" 
ale-py 
autorom 
pillow
opencv-python==4.10.0.84
numpy<2
pygame
matplotlib
wandb
python-dotenv
psycopg2-binary 
sqlalchemy